Refactor mtable, mtr into common and chtml components (mtd doesn't have common parts)
